Add files via upload
V1 - works for basic profiles with experience and education listed as one position per company. If a profile lists a company with multiple positions that experience is skipped over for the next company & position pairing that is 1:1. Working on a new version that will identify multiple jobs @ one company and properly pull that company and most recent position.